% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acknowledgements.R
\name{acknowledgements}
\alias{acknowledgements}
\title{acknowledgements}
\usage{
acknowledgements()
}
\value{
A list of people who have contributed something beyond code to the
ExPosition family of packages.
}
\description{
\code{acknowledgements} returns a list of people who have contributed to
ExPosition.
}
\author{
Derek Beaton
}
\keyword{misc}
