% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExPosition-package.R
\docType{data}
\name{authors}
\alias{authors}
\title{(A truncated form of) Punctuation used by six authors (data).}
\format{
authors$ca$data: Six authors (rows) and the frequency of three
puncutuations (columns). For use with \code{\link{epCA}}.\cr
authors$mca$data: A Burt table reformatting of the $ca$data. For use with
\code{\link{epMCA}}.
}
\usage{
data(authors)
}
\description{
How six authors use 3 different types of puncatuation throughout their
writing.
}
\references{
Brunet, E. (1989). Faut-il ponderer les donnees linguistiques.
\emph{CUMFID}, 16, 39-50. \cr Abdi, H., and Williams, L.J. (2010). Principal
component analysis. \emph{Wiley Interdisciplinary Reviews: Computational
Statistics}, 2, 433-459.\cr Abdi, H., and Williams, L.J. (2010).
Correspondence analysis. In N.J. Salkind, D.M., Dougherty, & B. Frey (Eds.):
\emph{Encyclopedia of Research Design}. Thousand Oaks (CA): Sage. pp.
267-278.
}
\keyword{datasets}
