\name{bada.wine}

\alias{bada.wine}

\docType{data}

\title{Twelve wines from 3 regions in France with 18 attributes.}

\description{
This data should be used for discriminant analyses or analyses where the \emph{group} information is important.
}

\usage{data(bada.wine)}

\format{
bada.wine$data: Data matrix with twelve wines (rows) from 3 regions with 18 attributes (columns).\cr
bada.wine$design: Design matrix with twelve wines (rows) with 3 regions (columns) to indicate group relationship of the data matrix.
}

\references{
Abdi, H. and Williams, L.J. (2010). Barycentric discriminant analysis (BADIA). In N.J. Salkind, D.M., Dougherty, & B. Frey (Eds.): \emph{Encyclopedia of Research Design}. Thousand Oaks (CA): Sage. pp. 64-75.
}

\keyword{datasets}
