% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExPosition-package.R
\docType{data}
\name{beer.tasting.notes}
\alias{beer.tasting.notes}
\title{Some of authors' personal beer tasting notes.}
\format{
beer.tasting.notes$data: Data matrix. Tasting notes (ratings) of 38
different beers (rows) described by 16 different flavor profiles
(columns).\cr beer.tasting.notes$brewery.design: Design matrix. Source
brewery of 38 different beers (rows) across 26 breweries (columns). \cr
beer.tasting.notes$style.design: Design matrix. Style of 38 different beers
(rows) across 20 styles (columns) (styles as listed from Beer Advocate
website). \cr beer.tasting.notes$sup.data: Supplementary data matrix. ABV
and overall preference ratings of 38 beers described by two features (ABV &
overall) in original value and rounded value.
}
\source{
Jenny Rieck and Derek Beaton laboriously ``collected'' these data
for ``experimental purposes''.
}
\usage{
data(beer.tasting.notes)
}
\description{
Tasting notes, preferences, breweries and styles of 38 different craft beers
from various breweries, across various styles.
}
\references{
http://www.beeradvocate.com
}
\keyword{datasets}
