% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExPosition-package.R
\docType{data}
\name{beers2007}
\alias{beers2007}
\title{Ten assessors sort eight beers into groups.}
\format{
beer2007$data: A data matrix with 8 rows (beers) described by 10
assessors (columns).
}
\usage{
data(beers2007)
}
\description{
Ten assessors perform a free-sorting task to sort eight beers into groups.
}
\references{
Abdi, H., Valentin, D., Chollet, S., & Chrea, C. (2007).
Analyzing assessors and products in sorting tasks: DISTATIS, theory and
applications. \emph{Food Quality and Preference}, 627-640.
}
\keyword{datasets}
