\name{caNorm}
\alias{caNorm}
\title{
Correspondence Analysis normalization.
}
\description{
Performs normalization for CA-type data.
}
\usage{
caNorm(X, X_dimensions, colTotal, rowTotal, grandTotal, weights = NULL, masses = NULL)
}
\arguments{
  \item{X}{}
  \item{X_dimensions}{}
  \item{colTotal}{}
  \item{rowTotal}{}
  \item{grandTotal}{}
  \item{weights}{}
  \item{masses}{}
}
\details{

}
\value{

}
\references{

}
\author{
Derek Beaton
}

\seealso{

}
\examples{
}
\keyword{ misc }
