\name{caSupplementalElementsPreProcessing}
\alias{caSupplementalElementsPreProcessing}
\title{
Correspondence Analysis preprocessing. 
}
\description{
CA preprocessing for data. Can be performed on rows or columns of your data. This is a row-profile normalization.
}
\usage{
caSupplementalElementsPreProcessing(SUP.DATA, hellinger = FALSE)
}
\arguments{
  \item{SUP.DATA}{}
  \item{hellinger}{}
}
\details{
}
\value{
}
\references{
}
\author{
Derek Beaton
}
\note{

}

\seealso{

}
\examples{

}
\keyword{ misc }
