\name{chi2Dist}
\alias{chi2Dist}
\title{
Chi-square Distance computation
}
\description{
Performs a chi-square distance. Primarily used for \code{\link{epMDS}}.
}
\usage{
chi2Dist(X)
}
\arguments{
  \item{X}{}
}
\details{
}
\value{

}
\references{

}
\author{
Hervé Abdi
}
\note{
}
\seealso{
}
\examples{
}

\keyword{ misc }
