% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDefaultDesign.R
\name{createDefaultDesign}
\alias{createDefaultDesign}
\title{createDefaultDesign}
\usage{
createDefaultDesign(DATA)
}
\arguments{
\item{DATA}{original data that requires a design matrix}
}
\value{
\item{DESIGN}{a column-vector matrix to indicate that all
observations are in the same group.}
}
\description{
Creates a default design matrix, wherein all observations (i.e., row items)
are in the same group.
}
\author{
Derek Beaton
}
\keyword{misc}
