\name{dica.ad}

\alias{dica.ad}

\docType{data}

\title{Alzheimer's Patient-Spouse Dyads.}

\description{
Conversational data from Alzheimer's Patient-Spouse Dyads.
}

\usage{data(dica.ad)}

\format{
dica.ad$data: Seventeen dyads described by 58 variables.\cr
dica.ad$design: Seventeen dyads that belong to three groups.
}

\references{
Williams, L.J., Abdi, H., French, R., & Orange, J.B. (2010). A tutorial on Multi-Block Discriminant Correspondence Analysis (MUDICA): A new method for analyzing discourse data from clinical populations. \emph{Journal of Speech Language and Hearing Research},\bold{ 53}, 1372-1393.
}

\keyword{datasets}
