\name{dica.wine}

\alias{dica.wine}

\docType{data}

\title{Twelve wines from 3 regions in France with 16 attributes.}

\description{
This data should be used for discriminant analyses or analyses where the \emph{group} information is important.
}

\usage{data(dica.wine)}

\format{
dica.wine$data: Data matrix with twelve wines (rows) from 3 regions with 16 attributes (columns) \emph{in disjunctive (0/1) coding}.\cr
dica.wine$design: Design matrix with twelve wines (rows) with 3 regions (columns) to indicate group relationship of the data matrix.
}

\references{
Abdi, H. (2007). Discriminant correspondence analysis. In N.J. Salkind (Ed.): \emph{Encyclopedia of Measurement and Statistics}. Thousand Oaks (CA): Sage. pp. 270-275.
}

\keyword{datasets}
