\name{ep.iris}

\alias{ep.iris}

\docType{data}

\title{Fisher's iris Set (for ExPosition)}

\description{
The world famous Fisher's iris set: 150 flowers from 3 species with 4 attributes.
}

\usage{data(ep.iris)}

\format{
ep.iris$data: Data matrix with 150 flowers (rows) from 3 species with 4 attributes (columns) describing sepal and petal features.\cr
ep.iris$design: Design matrix with 150 flowers (rows) with 3 species (columns) indicating which flower belongs to which species.
}

\source{
http://en.wikipedia.org/wiki/Iris_flower_data_set
}

\keyword{datasets}
