% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExPosition-package.R
\docType{data}
\name{faces2005}
\alias{faces2005}
\title{Faces analyzed using Four Algorithms}
\format{
faces2005$data: A data structure representing a distance matrix
(6X6) for four algorithms.
}
\usage{
data(faces2005)
}
\description{
Four algorithms compared using a distance matrix between six faces.
}
\references{
Abdi, H., & Valentin, D. (2007). DISTATIS: the analysis of
multiple distance matrices. \emph{Encyclopedia of Measurement and
Statistics}. 284-290.
}
\keyword{datasets}
