% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExPosition-package.R
\docType{data}
\name{french.social}
\alias{french.social}
\title{How twelve French families spend their income on groceries.}
\format{
french.social$data: Data matrix with twelve families (rows) with 7
attributes (columns) describing what they spend their income on.
}
\usage{
data(french.social)
}
\description{
This data should be used with \code{\link{epPCA}}
}
\references{
Lebart, L., and Fénelon, J.P. (1975) \emph{Statistique et
informatique appliquées}. Paris: Dunod\cr Abdi, H., and Williams, L.J.
(2010). Principal component analysis. \emph{Wiley Interdisciplinary Reviews:
Computational Statistics}, 2, 433-459.
}
\keyword{datasets}
