% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExPosition-package.R
\docType{data}
\name{great.beer.tasting.1}
\alias{great.beer.tasting.1}
\title{A collection of beer tasting notes from untrained assessors.}
\format{
great.beer.tasting.1$data: Data matrix (cube). Tasting notes
(ratings) of 9 different beers (rows) described by 16 different flavor
profiles (columns) by 4 untrained assessors. Thes data contain NAs and must
be imputed or adjusted before an analysis is performed.\cr
great.beer.tasting.1$brewery.design: Design matrix. Source brewery of 9
different beers (rows) across 5 breweries (columns). \cr
great.beer.tasting.1$flavor: Design matrix. Intended prominent flavor of 9
different beers (rows) across 3 flavor profiles (columns).
}
\source{
Rachel Williams, Jenny Rieck and Derek Beaton recoded, collected
data and/or ``ran the experiment''.
}
\usage{
data(great.beer.tasting.1)
}
\description{
A collection of beer tasting notes of 9 beers, across 16 descriptors, from 4
untrained assessors.
}
\keyword{datasets}
