\name{great.beer.tasting.2}
  
\alias{great.beer.tasting.2}

\docType{data}

\title{A collection of beer tasting notes from untrained assessors.}

\description{
A collection of beer tasting notes of 13 beers, across 15 descriptors, from 9 untrained assessors.
}

\usage{data(great.beer.tasting.2)}

\format{
great.beer.tasting.2$data: Data matrix (cube). Tasting notes (ratings) of 13 different beers (rows) described by 15 different flavor profiles (columns) by 9 untrained assessors. All original values were on an interval scale of 0-5. Any decimal values are imputed from alternate data sources or additional assessors.\cr
great.beer.tasting.2$brewery.design: Design matrix. Source brewery of 13 different beers (rows) across 13 breweries (columns). \cr
great.beer.tasting.2$style.design: Design matrix. Style of 13 different beers (rows) across 8 styles (columns). Some complex styles were truncated.
}


\source{
Rachel Williams, Jenny Rieck and Derek Beaton recoded, collected data and/or ``ran the experiment''.
}

\keyword{datasets}
