\name{jlsr.2010.ad}

\alias{jlsr.2010.ad}

\docType{data}

\title{Data from 17 Alzheimer's Patient-Spouse dyads.}

\description{
Seventeen Alzheimer's Patient-Spouse Dyads had conversations recorded and 58 attributes were recoded for this data. Each attribute is a frequency of occurence of the item.
}

\usage{data(jlsr.2010.ad)}

\format{
jlsr.2010.ad$ca$data: Seventeen patient-spouse dyads (rows) described by 58 conversation items. For use with \code{\link{epCA}} and discriminant analyses.\cr
jlsr.2010.ad$mca$design: A design matrix that indicates which group the dyad belongs to: control (CTRL), early stage Alzheimer's (EDAT) or middle stage Alzheimer's (MDAT).
}

\references{
Williams, L.J., Abdi, H., French, R., and Orange, J.B. (2010). A tutorial on Multi-Block Discriminant Correspondence Analysis (MUDICA): A new method for analyzing discourse data from clinical populations. \emph{Journal of Speech Language and Hearing Research}, 53, 1372-1393.
}

\keyword{datasets}
