\name{makeDistancesAndWeights}
\alias{makeDistancesAndWeights}
\title{
Makes distances and weights for MDS analyses (see \code{\link{epMDS}}).
}
\description{
	Makes distances and weights for MDS analyses (see \code{\link{epMDS}}).
}
\usage{
makeDistancesAndWeights(DATA, method = "euclidean", masses = NULL)
}
\arguments{
  \item{DATA}{}
  \item{method}{}
  \item{masses}{}
}
\details{
}
\value{
}
\references{
}
\author{
Derek Beaton
}
\note{
}

\seealso{

}
\examples{
}
\keyword{ misc }
