\name{mca.wine}

\alias{mca.wine}

\docType{data}

\title{Six wines described by several assessors with qualitative attributes.}

\description{
Six wines described by several assessors with qualitative attributes.
}

\usage{data(mca.wine)}

\format{
mca.wine$data: A (categorical) data matrix with 6 wines (rows) from several assessors described by 10 attributes (columns). For use with \code{\link{epMCA}}.
}

\references{
Abdi, H., & Valentin, D. (2007). Multiple correspondence analysis. In N.J. Salkind (Ed.): \emph{Encyclopedia of Measurement and Statistics}. Thousand Oaks (CA): Sage. pp. 651-657.
}

\keyword{datasets}
