\name{mdsSupplementalElementsPreProcessing}
\alias{mdsSupplementalElementsPreProcessing}
\title{
MDS preprocessing
}
\description{
Preprocessing of supplemental data for MDS analyses.
}
\usage{
mdsSupplementalElementsPreProcessing(SUP.DATA = NULL, D = NULL, M = NULL)
}
\arguments{
  \item{SUP.DATA}{}
  \item{D}{}
  \item{M}{}
}
\details{
}
\value{
}
\references{

}
\author{
Derek Beaton
}
\note{
}

\seealso{

}
\examples{

}
\keyword{ misc }
