% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pause.R
\name{pause}
\alias{pause}
\title{pause}
\usage{
pause(x = 0)
}
\arguments{
\item{x}{optional. If x>0 a call is made to \code{\link{Sys.sleep}}. Else,
execution pauses until a key is entered.}
}
\description{
A replication of MatLab pause function.
}
\references{
Copied from:\cr https://stat.ethz.ch/pipermail/r-help/2001-November/
}
\author{
Derek Beaton (but the pase of which is provided by Phillipe Brosjean
from the R mailing list.)
}
\keyword{misc}
