\name{pcaSupplementaryColsPreProcessing}
\alias{pcaSupplementaryColsPreProcessing}
\title{
Preprocessing for supplementary columns in PCA.
}
\description{
Preprocessing for supplementary columns in PCA.
}
\usage{
pcaSupplementaryColsPreProcessing(SUP.DATA = NULL, center = TRUE, scale = TRUE, M = NULL)
}
\arguments{
  \item{SUP.DATA}{}
  \item{center}{}
  \item{scale}{}
  \item{M}{}
}
\details{
}
\value{
}
\references{

}
\author{
Derek Beaton
}
\note{
}
\seealso{
}
\examples{
}
\keyword{ misc }
