\name{pcaSupplementaryRowsPreProcessing}
\alias{pcaSupplementaryRowsPreProcessing}
\title{
Preprocessing for rows in PCA.
}
\description{
Preprocessing for rows in PCA.
}
\usage{
pcaSupplementaryRowsPreProcessing(SUP.DATA = NULL, center = TRUE, scale = TRUE, W = NULL)
}
\arguments{
  \item{SUP.DATA}{}
  \item{center}{}
  \item{scale}{}
  \item{W}{}
}
\details{
}
\value{
}
\references{

}
\author{
Derek Beaton
}
\note{
}

\seealso{

}
\examples{
}
\keyword{ misc }
