\name{print.epCA}
\alias{print.epCA}
\title{Print Correspondence Analysis (CA) results}
\description{
	Print Correspondence Analysis (CA) results.
}
\usage{
	\method{print}{epCA}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the epCA class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton and Cherise Chin-Fatt}
\keyword{print}
