\name{print.epGPCA}
\alias{print.epGPCA}
\title{Print Generalized Principal Components Analysis (GPCA) results}
\description{
	Print Generalized Principal Components Analysis (GPCA) results.
}
\usage{
	\method{print}{epGPCA}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the epGPCA class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton and Cherise Chin-Fatt}
\keyword{print}
