% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.epGraphs.R
\name{print.epGraphs}
\alias{print.epGraphs}
\title{Print epGraphs results}
\usage{
\method{print}{epGraphs}(x,\dots)
}
\arguments{
\item{x}{an list that contains items to make into the epGraphs class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print epGraphs results
}
\seealso{
\code{\link{epGraphs}}
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
