% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.epMDS.R
\name{print.epMDS}
\alias{print.epMDS}
\title{Print Multidimensional Scaling (MDS) results}
\usage{
\method{print}{epMDS}(x,\dots)
}
\arguments{
\item{x}{an list that contains items to make into the epMDS class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print Multidimensional Scaling (MDS) results
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
