% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.epPCA.R
\name{print.epPCA}
\alias{print.epPCA}
\title{Print Principal Components Analysis (PCA) results}
\usage{
\method{print}{epPCA}(x,\dots)
}
\arguments{
\item{x}{an list that contains items to make into the epPCA class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print Principal Components Analysis (PCA) results
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
