% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.epSVD.R
\name{print.epSVD}
\alias{print.epSVD}
\title{Print results from the singular value decomposition (SVD) in ExPosition}
\usage{
\method{print}{epSVD}(x,\dots)
}
\arguments{
\item{x}{an list that contains items to make into the epSVD class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print results from the singular value decomposition (SVD) in ExPosition
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
