% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.expoOutput.R
\name{print.expoOutput}
\alias{print.expoOutput}
\title{Print results from ExPosition}
\usage{
\method{print}{expoOutput}(x,\dots)
}
\arguments{
\item{x}{an list that contains items to make into the expoOutput class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print results from ExPosition
}
\seealso{
\code{\link{epPCA}}, \code{\link{epGraphs}}
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
