\name{rvCoeff}
\alias{rvCoeff}
\title{
Perform Rv coefficient computation.
}
\description{
Perform Rv coefficient computation.
}
\usage{
rvCoeff(S, T, type = -1)
}
\arguments{
  \item{S}{}
  \item{T}{}
  \item{type}{}
}
\details{
}
\value{
}
\references{

}
\author{
Derek Beaton
}
\note{}
\seealso{
}
\examples{

}
\keyword{ misc }
