% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt.mat.R
\name{sqrt.mat}
\alias{sqrt.mat}
\title{sqrt.mat}
\usage{
\method{sqrt}{mat}(X)
}
\arguments{
\item{X}{a matrix that is square and symmetric}
}
\description{
sqrt.mat performs the square root of a matrix only for square symmetric matrices
This function should not be used directly.
}
\author{
Derek Beaton
}
