% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt.mat.R
\name{sqrt_mat}
\alias{sqrt_mat}
\title{sqrt_mat}
\usage{
sqrt_mat(X)
}
\arguments{
\item{X}{a matrix that is square and symmetric}
}
\description{
sqrt_mat performs the square root of a matrix only for square symmetric matrices
This function should not be used directly.
}
\author{
Derek Beaton
}
