% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExPosition-package.R
\docType{data}
\name{wines2007}
\alias{wines2007}
\title{Six wines described by 3 assessors.}
\format{
wines2007$data: A data set with 3 experts (studies) describing 6
wines (rows) using several variables using a scale from 1 to 7 with a total
of 10 measures (columns).\cr wines2007$table: A data matrix which identifies
the 3 experts (studies). \cr
}
\usage{
data(wines2007)
}
\description{
How six wines are described by 3 assessors across various flavor profiles,
totaling 10 columns.
}
\references{
Abdi, H., & Valentin, D. (2007). STATIS. In N.J. Salkind (Ed.):
\emph{Encyclopedia of Measurement and Statistics}. Thousand Oaks (CA): Sage.
pp. 955-962.
}
\keyword{datasets}
