\name{wines2012}
  
\alias{wines2012}

\docType{data}

\title{Wines Data from 12 assessors described by 15 flavor profiles.}

\description{
10 experts who describe 12 wines using four variables (cat-pee, passion fruit, green pepper, and mineral) considered as standard, and up to two additional variables if the experts chose. 
}

\usage{data(wines2012)}

\format{
wines2012$data: A data set with 10 experts (studies) describing 12 wines (rows) using four to six variables using a scale from 1 to 9 with a total of 53 measures (columns).\cr
wines2012$table: A data matrix which identifies the 10 experts (studies). \cr
wines2012$supplementary: A data matrix with 12 wines (rows) describing 4 Chemical Properties (columns).\cr
}

\references{
Abdi, H., Williams, L.J., Valentin, D., & Bennani-Dosse, M. (2012). STATIS and DISTATIS: Optimum multi-table principal component analysis and three way metric multidimensional scaling. \emph{Wiley Interdisciplinary Reviews: Computational Statistics}, 4, 124-167.
}

\keyword{datasets}
