\name{words}

\alias{words}

\docType{data}

\title{Twenty words described by 2 features.}

\description{
Twenty words ``randomly'' selected from a dictionary and described by two features: length of word and number of definitions.
}

\usage{data(words)}

\format{
words$data: A data matrix with 20 words (rows) described by 2 attributes (columns). For use with \code{\link{epPCA}}.
}

\references{
Abdi, H., and Williams, L.J. (2010). Principal component analysis. \emph{Wiley Interdisciplinary Reviews: Computational Statistics}, 2, 433-459.
}

\keyword{datasets}
