\name{acknowledgements}
\alias{acknowledgements}
\title{
acknowledgements
}
\description{
\code{acknowledgements} returns a list of people who have contributed to ExPosition.
}
\usage{
acknowledgements()
}

\value{
	A list of people who have contributed something beyond code to the ExPosition family of packages.
}
\author{
Derek Beaton
}
\keyword{ misc }