% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExPosition-package.R
\docType{data}
\name{bada.wine}
\alias{bada.wine}
\title{Twelve wines from 3 regions in France with 18 attributes.}
\format{
bada.wine$data: Data matrix with twelve wines (rows) from 3 regions
with 18 attributes (columns).\cr bada.wine$design: Design matrix with twelve
wines (rows) with 3 regions (columns) to indicate group relationship of the
data matrix.
}
\usage{
data(bada.wine)
}
\description{
This data should be used for discriminant analyses or analyses where the
\emph{group} information is important.
}
\references{
Abdi, H. and Williams, L.J. (2010). Barycentric discriminant
analysis (BADIA). In N.J. Salkind, D.M., Dougherty, & B. Frey (Eds.):
\emph{Encyclopedia of Research Design}. Thousand Oaks (CA): Sage. pp. 64-75.
}
\keyword{datasets}
