% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2Dist.R
\name{chi2Dist}
\alias{chi2Dist}
\title{Chi-square Distance computation}
\usage{
chi2Dist(X)
}
\arguments{
\item{X}{Compute chi-square distances between row items.}
}
\value{
\item{D}{Distance matrix for \code{\link{epMDS}} analysis.}
\item{MW}{a list of masses and weights. Weights not used in MDS.}
}
\description{
Performs a chi-square distance. Primarily used for \code{\link{epMDS}}.
}
\author{
Hervé Abdi
}
\keyword{misc}
\keyword{multivariate}
