\name{coffee.data}

\alias{coffee.data}

\docType{data}

\title{Small data set on flavor perception and preferences for coffee.}

\description{
One coffee from Oak Cliff roasters (Dallas, TX) was used in this experiment. Honduran source with a medium roast. The coffee was brewed in two ways and served in two ways (i.e., a 2x2 design). Two batches each of coffee were brewed at 180 degrees (Hot) Farenheit or at room temperature (Cold). One of each was served cold or heated back up to 180 degrees (Hot).
}

\details{
	Flavor profiles measured: Salty, Spice Cabinet, Sweet, Bittery, and Nutty.
}

\usage{data(coffee.data)}

\format{
coffee.data$preferences: Ten participants indicated if they liked a particular serving or not.\cr
coffee.data$ratings: Ten participants indicated on a scale of 0-2 the presence of particular flavors. In an array format.
}

\keyword{datasets}
