% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMW.R
\name{computeMW}
\alias{computeMW}
\title{computeMW}
\usage{
computeMW(DATA, masses = NULL, weights = NULL)
}
\arguments{
\item{DATA}{original data; will be used to compute masses and weights if
none are provided.}

\item{masses}{a vector or (diagonal) matrix of masses for the row items. If
NULL (default), masses are computed as 1/# of rows}

\item{weights}{a vector or (diagonal) matrix of weights for the column
items. If NULL (default), weights are computed as 1/# of columns}
}
\value{
Returns a list with the following items:\cr \item{M}{a diagonal
matrix of masses (if too large, a vector is returned).} \item{W}{a diagonal
matrix of weights (if too large, a vector is returned).}
}
\description{
Computes masses and weights for use.
}
\author{
Derek Beaton
}
\keyword{misc}
\keyword{multivariate}
