% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExPosition-package.R
\docType{data}
\name{dica.wine}
\alias{dica.wine}
\title{Twelve wines from 3 regions in France with 16 attributes.}
\format{
dica.wine$data: Data matrix with twelve wines (rows) from 3 regions
with 16 attributes (columns) \emph{in disjunctive (0/1) coding}.\cr
dica.wine$design: Design matrix with twelve wines (rows) with 3 regions
(columns) to indicate group relationship of the data matrix.
}
\usage{
data(dica.wine)
}
\description{
This data should be used for discriminant analyses or analyses where the
\emph{group} information is important.
}
\references{
Abdi, H. (2007). Discriminant correspondence analysis. In N.J.
Salkind (Ed.): \emph{Encyclopedia of Measurement and Statistics}. Thousand
Oaks (CA): Sage. pp. 270-275.
}
\keyword{datasets}
