\name{faces2005}

\alias{faces2005}

\docType{data}

\title{Faces analyzed using Four Algorithms}

\description{
Four algorithms compared using a distance matrix between six faces. 
}

\usage{data(faces2005)}

\format{
faces2005$data: A data structure representing a distance matrix (6X6) for four algorithms.
}

\references{
Abdi, H., & Valentin, D. (2007). DISTATIS: the analysis of multiple distance matrices. \emph{Encyclopedia of Measurement and Statistics}. 284-290.
}

\keyword{datasets}
