\name{french.social}
\alias{french.social}
\encoding{UTF-8}
\docType{data}

\title{How twelve French families spend their income on groceries.}

\description{
This data should be used with \code{\link{epPCA}}
}

\usage{data(french.social)}

\format{
french.social$data: Data matrix with twelve families (rows) with 7 attributes (columns) describing what they spend their income on.
}

\references{
Lebart, L., and Fénelon, J.P. (1975) \emph{Statistique et informatique appliquées}. Paris: Dunod\cr
Abdi, H., and Williams, L.J. (2010). Principal component analysis. \emph{Wiley Interdisciplinary Reviews: Computational Statistics}, 2, 433-459.
}

\keyword{datasets}
