\name{jocn.2005.fmri}

\alias{jocn.2005.fmri}

\docType{data}

\title{Data of categories of images as view in an \emph{f}MRI experiment.}

\description{
Contains 2 data sets: distance matrix of \emph{f}MRI scans of participants viewing categories of items and distance matrix of the actual pixels from the images in each category.
}

\usage{data(jocn.2005.fmri)}

\format{
jocn.2005.fmri$images$data: A distance matrix of 6 categories of images based on a pixel analysis. \cr
jocn.2005.fmri$scans$data: A distance matrix of 6 categories of images based on \emph{f}MRI scans.
}

\references{
O'Toole, A. J., Jiang, F., Abdi, H., and Haxby, J. V. (2005). Partially distributed representations of objects and faces in ventral temporal cortex. \emph{Journal of Cognitive Neuroscience}, \emph{17}(4), 580-590. \cr
Haxby, J. V., Gobbini, M. I., Furey, M. L., Ishai, A., Schouten, J. L., and Pietrini, P. (2001). Distributed and overlapping representation of faces and objects in ventral temporal cortex. \emph{Science}, 293, 2425-2430.\cr
}

\seealso{
	http://openfmri.org/dataset/ds000105
}

\keyword{datasets}
