% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNominalData.R
\name{makeNominalData}
\alias{makeNominalData}
\title{makeNominalData}
\usage{
makeNominalData(datain)
}
\arguments{
\item{datain}{a data matrix where the \emph{columns} will be recoded.}
}
\value{
\item{dataout}{a transformed version of \emph{datain}.}
}
\description{
Transforms each column into measure-response columns with disjunctive (0/1)
coding. If NA is found somewhere in matrix, barycentric recoding is peformed
for the missing value(s).
}
\examples{

	data(mca.wine)
	nominal.wine <- makeNominalData(mca.wine$data)

}
\seealso{
\code{\link{epMCA}}
}
\author{
Derek Beaton
}
\keyword{misc}
\keyword{multivariate}
