% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExPosition-package.R
\docType{data}
\name{mca.wine}
\alias{mca.wine}
\title{Six wines described by several assessors with qualitative attributes.}
\format{
mca.wine$data: A (categorical) data matrix with 6 wines (rows) from
several assessors described by 10 attributes (columns). For use with
\code{\link{epMCA}}.
}
\usage{
data(mca.wine)
}
\description{
Six wines described by several assessors with qualitative attributes.
}
\references{
Abdi, H., & Valentin, D. (2007). Multiple correspondence
analysis. In N.J. Salkind (Ed.): \emph{Encyclopedia of Measurement and
Statistics}. Thousand Oaks (CA): Sage. pp. 651-657.
}
\keyword{datasets}
