\name{mdsTransform}
\alias{mdsTransform}
\title{
Transform data for MDS analysis.
}
\description{
Transform data for MDS analysis.
}
\usage{
mdsTransform(D, MW)
}
\arguments{
  \item{D}{}
  \item{MW}{}
}
\details{
}
\value{
}
\references{
}
\author{
Derek Beaton
}
\note{
}
\seealso{
}
\examples{
}
\keyword{ misc }
