\name{pause}
\alias{pause}
\title{
pause
}
\description{
A replication of MatLab pause function.
}
\usage{
pause(x = 0)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{optional. If x>0 a call is made to \code{\link{Sys.sleep}}. Else, execution pauses until a key is entered.}
}

\references{
 
Copied from:\cr
https://stat.ethz.ch/pipermail/r-help/2001-November/
}
\author{
Derek Beaton (but the pase of which is provided by Phillipe Brosjean from the R mailing list.)
}

\keyword{ misc }
