\name{pca.wine}

\alias{pca.wine}

\docType{data}

\title{Six wines described by several assessors with rank attributes.}

\description{
Six wines described by several assessors with rank attributes.
}

\usage{data(pca.wine)}

\format{
pca.wine$data: A data matrix with 6 wines (rows) from several assessors described by 11 attributes (columns). For use with \code{\link{epPCA}}.
}

\references{
Abdi, H., and Williams, L.J. (2010). Principal component analysis. \emph{Wiley Interdisciplinary Reviews: Computational Statistics}, 2, 433-459.
}

\seealso{
	\link{mca.wine}
}

\keyword{datasets}
