% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaSupplementaryColsPreProcessing.R
\name{pcaSupplementaryColsPreProcessing}
\alias{pcaSupplementaryColsPreProcessing}
\title{Preprocessing for supplementary columns in PCA.}
\usage{
pcaSupplementaryColsPreProcessing(SUP.DATA = NULL, center = TRUE,
scale = TRUE, M = NULL)
}
\arguments{
\item{SUP.DATA}{A supplemental matrix that has the same number of rows as an
active data set.}

\item{center}{The center from the active data. NULL will center
\code{SUP.DATA} to itself.}

\item{scale}{The scale factor from the active data. NULL will scale
(z-score) \code{SUP.DATA} to itself.}

\item{M}{Masses from the active data.}
}
\value{
a matrix that has been preprocessed to project supplementary columns
for PCA methods.
}
\description{
Preprocessing for supplementary columns in PCA.
}
\author{
Derek Beaton
}
\keyword{misc}
\keyword{multivariate}
