% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.epCA.R
\name{print.epCA}
\alias{print.epCA}
\title{Print Correspondence Analysis (CA) results}
\usage{
\method{print}{epCA}(x,\dots)
}
\arguments{
\item{x}{an list that contains items to make into the epCA class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print Correspondence Analysis (CA) results
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
