% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.epMCA.R
\name{print.epMCA}
\alias{print.epMCA}
\title{Print Multiple Correspondence Analysis (MCA) results}
\usage{
\method{print}{epMCA}(x,\dots)
}
\arguments{
\item{x}{an list that contains items to make into the epMCA class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print Multiple Correspondence Analysis (MCA) results
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
