\name{print.epMDS}
\alias{print.epMDS}
\title{Print Multidimensional Scaling (MDS) results}
\description{
	Print Multidimensional Scaling (MDS) results.
}
\usage{
	\method{print}{epMDS}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the epMDS class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton and Cherise Chin-Fatt}
\keyword{print}
