\name{print.epPCA}
\alias{print.epPCA}
\title{Print Principal Components Analysis (PCA) results}
\description{
	Print Principal Components Analysis (PCA) results.
}
\usage{
	\method{print}{epPCA}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the epPCA class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton and Cherise Chin-Fatt}
\keyword{print}
