\name{print.expoOutput}
\alias{print.expoOutput}
\title{Print results from ExPosition}
\description{
	Print results from the ExPosition. Includes results from a given method and epGraphs.
}
\usage{
	\method{print}{expoOutput}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the expoOutput class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\seealso{
	\code{\link{epPCA}}, \code{\link{epGraphs}}
}
\author{Derek Beaton and Cherise Chin-Fatt}
\keyword{print}
