\name{rvCoeff}
\alias{rvCoeff}
\encoding{UTF-8}
\title{
Perform Rv coefficient computation.
}
\description{
Perform Rv coefficient computation.
}
\usage{
rvCoeff(Smat, Tmat)
}
\arguments{
  \item{Smat}{A square covariance matrix}
  \item{Tmat}{A square covariance matrix}
}
\value{
	A single value that is the Rv coefficient. 
}
\references{
	Robert, P., & Escoufier, Y. (1976). A Unifying Tool for Linear Multivariate Statistical Methods: The RV-Coefficient. \emph{Journal of the Royal Statistical Society. Series C (Applied Statistics)}, \emph{25}(3), 257--265.}
\author{
Derek Beaton
}
\keyword{ misc }
\keyword{ multivariate }
