\name{supplementalProjection}
\alias{supplementalProjection}
\title{
Supplemental projections.
}
\description{
Performs a supplementary projection across ExPosition (and related) techniques.
}
\usage{
supplementalProjection(sup.transform = NULL, f.scores = NULL, Dv = NULL)
}
\arguments{
  \item{sup.transform}{}
  \item{f.scores}{}
  \item{Dv}{}
}
\details{
}
\value{
}
\references{
}
\author{
Derek Beaton
}
\note{
}
\seealso{
}
\examples{

}
\keyword{ misc }